refactor(member) 파일 위치 추가. Service 형식 변경
