+++ member-service/src/main/java/com/green/member/application/admin/AdminController.java
@@ -1,8 +1,6 @@
 package com.green.member.application.admin;
 
 import com.green.common.model.ResultResponse;
-import com.green.member.application.member.MemberService;
-import com.green.member.application.member.model.MemberCreateReq;
 import com.green.member.application.member.model.MemberCreateRes;
 import com.green.member.application.student.model.StudentCreateReq;
 import lombok.RequiredArgsConstructor;
@@ -15,12 +13,12 @@ import org.springframework.web.multipart.MultipartFile;
 @RequiredArgsConstructor
 @RequestMapping("/admin")
 public class AdminController {
-    private final MemberService memberService;
+    private final AdminService adminService;
 
     @PostMapping("/students")
     public ResultResponse<?> createStudent(@RequestPart StudentCreateReq req,
                                            @RequestPart(required = false) MultipartFile pic) {
-        MemberCreateRes res = memberService.createStudent(req, pic);
+        MemberCreateRes res = adminService.createStudent(req, pic);
         return ResultResponse.builder()
                 .message("학생 정보 등록 성공")
                 .data(res)
