+++ member-service/src/main/java/com/green/member/application/member/MemberRepository.java
@@ -13,20 +13,19 @@ public interface MemberRepository  extends JpaRepository<Member, Long> {
         WHERE YEAR(m.entryDate) = :entryYear
     """)
     int countStudentByEntryYear(@Param("entryYear") int entryYear);
-//
-//    @Query("""
-//        SELECT COUNT(s) + 1
-//        FROM Professor p
-//        JOIN p.member m
-//        WHERE YEAR(m.entryDate) = :entryYear
-//    """)
-//    int countProfessorByEntryYear(@Param("entryYear") int entryYear);
-//
-//    @Query("""
-//        SELECT COUNT(s) + 1
-//        FROM Admin a
-//        JOIN a.member m
-//        WHERE YEAR(m.entryDate) = :entryYear
-//    """)
-//    int countAdminByEntryYear(@Param("entryYear") int entryYear);
+    @Query("""
+    SELECT COUNT(p) + 1
+    FROM Professor p
+    JOIN p.member m
+    WHERE YEAR(m.entryDate) = :entryYear
+""")
+    int countProfessorByEntryYear(@Param("entryYear") int entryYear);
+
+    @Query("""
+    SELECT COUNT(a) + 1
+    FROM Admin a
+    JOIN a.member m
+    WHERE YEAR(m.entryDate) = :entryYear
+""")
+    int countAdminByEntryYear(@Param("entryYear") int entryYear);
 }
